+++ app/composables/map/index.ts
@@ -68,10 +68,12 @@ export const useHeaderMenu = () => computed<HeaderItem[]>(() => {
             text: 'Dashboard',
             icon: DataIcon,
             path: '/dashboard',
-            children: seenDashboards.value.map(x => ({
-                text: x.name,
-                path: `/dashboard/${ x.id }`,
-            })),
+            children: seenDashboards.value.length
+                ? seenDashboards.value.map(x => ({
+                    text: x.name,
+                    path: `/dashboard/${ x.id }`,
+                }))
+                : undefined,
         },
         {
             text: 'Stats',