+++ app/store/map.ts
@@ -345,7 +345,7 @@ export const useMapStore = defineStore('map', {
                 this.openingOverlay = false;
 
                 overlay.data.airport = await $fetch<VatsimAirportData>(`/api/data/vatsim/airport/${ airport }`, {
-                    timeout: 5000,
+                    timeout: 15000,
                 });
                 $fetch<VatsimAirportDataNotam[]>(`/api/data/vatsim/airport/${ airport }/notams`).then(x => overlay.data.notams = x).catch(e => {
                     console.error(e);