+++ app/utils/settings/types.ts
@@ -112,6 +112,7 @@ export interface UserSettingsV2 {
 
         traffic: {
             showFullRoute: boolean;
+            showRouteDetails: boolean;
             toggleAircraftOverlays: boolean;
             autoShowAirportTracks: boolean;
             disableFastUpdate: boolean;